requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.4
django-crispy-forms==2.3
django-recaptcha==4.0.0
h11==0.16.0
html5lib==1.1
idna==3.10
lxml==5.4.0
oscrypto==1.3.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2024.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.3.0
webencodings==0.5.1
wsproto==1.2.0
xhtml2pdf==0.2.17

djangorestframework==3.15.1
gunicorn>=20.1.0
django-filter==24.2
whitenoise==6.6.0

# API Authentication and JWT
djangorestframework-simplejwt==5.3.0
dj-rest-auth==6.0.0
django-allauth==65.2.0

# API Documentation
drf-spectacular==0.27.2

# Additional API dependencies
inflection==0.5.1
jsonschema==4.23.0
django-cors-headers==4.7.0
